+++ plone/app/event/tests/base_setup.py
@@ -2,7 +2,6 @@ from Products.CMFCore.utils import getToolByName
 from datetime import datetime
 from datetime import timedelta
 from plone.app.event.dx import behaviors
-from plone.app.event.dx.behaviors import data_postprocessing_context
 from plone.app.event.testing import set_browserlayer
 from plone.app.event.testing import set_timezone
 from plone.app.testing import TEST_USER_ID
@@ -76,7 +75,6 @@ class AbstractSampleDataEvents(unittest.TestCase):
             recurrence='RRULE:FREQ=DAILY;COUNT=3')
         workflow.doActionFor(self.past_event, 'publish')
         # adjust start and end according to whole_day and open_end
-        data_postprocessing_context(self.past_event)
         self.past_event.reindexObject()
 
         self.now_event = factory(
@@ -98,7 +96,6 @@ EXDATE:20130506T000000,20140404T000000""",
         # https://github.com/plone/plone.dexterity/pull/18
         # https://github.com/plone/plone.app.dexterity/issues/118
         workflow.doActionFor(self.now_event, 'publish')
-        data_postprocessing_context(self.now_event)
         self.now_event.reindexObject()
 
         self.future_event = factory(
@@ -109,7 +106,6 @@ EXDATE:20130506T000000,20140404T000000""",
             end=future + duration,
             location=u'Graz')
         workflow.doActionFor(self.future_event, 'publish')
-        data_postprocessing_context(self.future_event)
         self.future_event.reindexObject()
 
         self.portal.invokeFactory('Folder', 'sub', title=u'sub')
@@ -121,7 +117,6 @@ EXDATE:20130506T000000,20140404T000000""",
             end=far,
             location=u'Schaftal')
         workflow.doActionFor(self.long_event, 'publish')
-        data_postprocessing_context(self.long_event)
         self.long_event.reindexObject()
 
         # For AT based tests, this is a plone.app.collection ICollection type