+++ CHANGES.rst
@@ -1,11 +1,19 @@
 Changelog
 =========
 
-2.0.0 (unreleased)
+2.1.0 (unreleased)
 ------------------
 
 - Some Plone 5 related js improvements
   [vangheem]
+
+- Use Plone 5 imports from plone.app.z3cform and make plone.app.widgets a soft
+  dependency.
+  [vangheem]
+
+
+2.0.0 (unreleased)
+------------------
 
 - Remove support for ``plone.app.collection`` and ``ATTopic`` - plone.app.event
   2.x is Dexterity only.